Fixed polyline feature with coordinate attribute Float support for ARC_DEGREES
git-svn-id: https://svn.osgeo.org/gdal/trunk@9636 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/ili/ogrili1datasource.cpp
@@ -28,6 +28,10 @@
  ******************************************************************************
  *
  * $Log$
+ * Revision 1.7  2006/04/24 16:49:48  pka
+ * Fixed polyline feature with coordinate attribute
+ * Float support for ARC_DEGREES
+ *
  * Revision 1.6  2005/12/20 16:47:04  pka
  * Interlis 1 output without model
  *
@@ -170,7 +174,7 @@ int OGRILI1DataSource::Open( const char * pszNewName, int bTestOpen )
 
     if( getenv( "ARC_DEGREES" ) != NULL ) {
       //No better way to pass arguments to the reader (it could even be an -lco arg)
-      poReader->SetArcDegrees( atoi( getenv("ARC_DEGREES") ) );
+      poReader->SetArcDegrees( atof( getenv("ARC_DEGREES") ) );
     }
 
     poReader->ReadFeatures(); //FIXME
